refactor: 피드 카드 공용 컴포넌트 분리 및 홈 UI 개선
