+++ SNAPY_iOS/Feature/Profile/Component/ProfileFeedGrid.swift
@@ -108,7 +108,7 @@ struct FeedDetailView: View {
     }
 }
 
-// MARK: - 피드 상세 카드 (단일 게시물)
+// MARK: - 피드 상세 카드
 struct FeedDetailCard: View {
     let post: FeedPost
     let displayName: String
@@ -116,216 +116,30 @@ struct FeedDetailCard: View {
     let profileImage: UIImage?
     let profileAsset: String
 
-    @State private var currentPage = 0
     @State private var isLiked = false
     @State private var likeCount = 0
-    @State private var commentCount = 0
-    @State private var showComments = false
-    @State private var heartAnimations: [HeartAnimation] = []
-    @State private var heartTapCount: Int = 0
 
     var body: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            // 프로필 헤더
-            HStack(spacing: 14) {
-                profileImageView
-                    .frame(width: 36, height: 36)
-                    .clipShape(Circle())
-                    .padding(3)
-                    .overlay(
-                        Circle()
-                            .stroke(Color.customGray500, lineWidth: 0.7)
-                    )
-
-                VStack(alignment: .leading, spacing: 4) {
-                    Text(displayName)
-                        .font(.system(size: 14, weight: .bold))
-                        .foregroundColor(.white)
-                    Text(handle)
-                        .font(.system(size: 12))
-                        .foregroundColor(.customGray300)
-                }
-
-                Spacer()
-
-                Text(post.date)
-                    .font(.system(size: 13))
-                    .foregroundColor(.customGray300)
-            }
-            .padding(.horizontal, 14)
-            .padding(.vertical, 10)
-
-            // 사진 슬라이더 (back 배경 + front 드래그 PIP + 더블탭 하트)
-            ZStack {
-                TabView(selection: $currentPage) {
-                    ForEach(Array(post.photos.enumerated()), id: \.offset) { index, photo in
-                        GeometryReader { geo in
-                            ZStack(alignment: .topLeading) {
-                                // back 이미지
-                                if let backUrl = photo.backImageUrl, let url = URL(string: backUrl) {
-                                    KFImage(url)
-                                        .resizable()
-                                        .placeholder { Color(white: 0.15) }
-                                        .fade(duration: 0.2)
-                                        .scaledToFill()
-                                        .frame(width: geo.size.width, height: geo.size.height)
-                                        .clipped()
-                                } else {
-                                    Color(white: 0.15)
-                                }
-
-                                // front 드래그 가능 PIP
-                                if let frontUrl = photo.frontImageUrl, let url = URL(string: frontUrl) {
-                                    DraggablePIP(
-                                        containerSize: geo.size,
-                                        pipWidth: 120,
-                                        pipHeight: 160,
-                                        padding: 12
-                                    ) {
-                                        KFImage(url)
-                                            .resizable()
-                                            .placeholder { Color(white: 0.2) }
-                                            .fade(duration: 0.2)
-                                            .scaledToFill()
-                                    }
-                                }
-                            }
-                        }
-                        .tag(index)
-                    }
-                }
-                .tabViewStyle(.page(indexDisplayMode: .never))
-
-                // 더블탭 하트 애니메이션
-                ForEach(heartAnimations) { heart in
-                    Image("Heart_img")
-                        .resizable()
-                        .scaledToFit()
-                        .frame(width: heart.size, height: heart.size)
-                        .rotationEffect(.degrees(heart.rotation))
-                        .scaleEffect(heart.scale)
-                        .opacity(heart.opacity)
-                        .position(heart.position)
-                }
-            }
-            .frame(height: 540)
-            .contentShape(Rectangle())
-            .onTapGesture(count: 2) { location in
-                triggerHeartAnimation(at: location)
-                if !isLiked {
-                    isLiked = true
-                    likeCount += 1
-                }
-            }
-
-            // 페이지 인디케이터 (1개여도 간격 유지)
-            HStack(spacing: 5) {
-                if post.photos.count > 1 {
-                    ForEach(0..<post.photos.count, id: \.self) { index in
-                        Circle()
-                            .fill(index == currentPage ? Color.MainYellow : Color.customGray300)
-                            .frame(width: 6, height: 6)
-                    }
-                }
-            }
-            .frame(maxWidth: .infinity)
-            .frame(height: 6)
-            .padding(.vertical, 14)
-
-            // 액션 버튼
-            HStack(spacing: 14) {
-                HStack(spacing: 5) {
-                    Button {
-                        isLiked.toggle()
-                        likeCount += isLiked ? 1 : -1
-                    } label: {
-                        Image(systemName: isLiked ? "heart.fill" : "heart")
-                            .font(.system(size: 26))
-                            .foregroundColor(isLiked ? .red : .white)
-                    }
-                    Text("\(likeCount)")
-                        .font(.system(size: 16, weight: .medium))
-                        .foregroundColor(.white)
-                }
-
-                HStack(spacing: 8) {
-                    Button {
-                        showComments = true
-                    } label: {
-                        Image(systemName: "bubble.right")
-                            .font(.system(size: 24))
-                            .foregroundColor(.white)
-                    }
-                    Text("\(commentCount)")
-                        .font(.system(size: 16, weight: .medium))
-                        .foregroundColor(.white)
-                }
-
-                Button {
-                    // 공유
-                } label: {
-                    Image(systemName: "paperplane")
-                        .font(.system(size: 20))
-                        .foregroundColor(.white)
-                }
-
-                Spacer()
-            }
-            .padding(.horizontal, 14)
-            .padding(.bottom, 20)
-
-            // 이미지 댓글 영역
-            ImageCommentSection()
-                .padding(.horizontal, 14)
-                .padding(.bottom, 12)
-        }
-        .sheet(isPresented: $showComments) {
-            CommentSheetView(postId: UUID())
-                .presentationDetents([.large])
-                .presentationDragIndicator(.hidden)
-        }
+        FeedCardView(
+            profileImageSource: profileSource,
+            displayName: displayName,
+            handle: handle,
+            date: post.date,
+            photos: post.photos.map {
+                FeedCardPhoto(frontImageUrl: $0.frontImageUrl, backImageUrl: $0.backImageUrl, assetName: nil)
+            },
+            isLiked: $isLiked,
+            likeCount: $likeCount
+        )
     }
 
-    // MARK: - 더블탭 하트
-
-    private func triggerHeartAnimation(at location: CGPoint) {
-        heartTapCount += 1
-        let size: CGFloat = 60 + CGFloat(heartTapCount - 1) * 2
-        let heart = HeartAnimation(position: location, size: min(size, 120))
-        heartAnimations.append(heart)
-
-        withAnimation(.spring(response: 0.3, dampingFraction: 0.5)) {
-            if let idx = heartAnimations.firstIndex(where: { $0.id == heart.id }) {
-                heartAnimations[idx].scale = 1.2
-                heartAnimations[idx].opacity = 1.0
-            }
-        }
-
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.6) {
-            withAnimation(.easeOut(duration: 0.3)) {
-                if let idx = heartAnimations.firstIndex(where: { $0.id == heart.id }) {
-                    heartAnimations[idx].scale = 1.6
-                    heartAnimations[idx].opacity = 0
-                }
-            }
-        }
-
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
-            heartAnimations.removeAll { $0.id == heart.id }
-        }
-    }
-
-    @ViewBuilder
-    private var profileImageView: some View {
+    private var profileSource: ProfileImageSource {
         if let image = profileImage {
-            Image(uiImage: image)
-                .resizable()
-                .scaledToFill()
-        } else {
-            Image(profileAsset)
-                .resizable()
-                .scaledToFill()
+            return .uiImage(image)
+        } else if !profileAsset.isEmpty {
+            return .asset(profileAsset)
         }
+        return .none
     }
 }
